Augment vendor scripts to be one-at-a-time oriented
1. Always reimport Cargo.lock from vendor tree before running, to ensure
   lock-delta is relative to the last known good state

2. Backup local Cargo.toml to vendor/ to give some concrete reference
   point for comparison as to what Cargo.toml was valid at the time of
   vendorization, on the basis that if you check-out some arbitrary
   commit in vendor/, as long as it matches Cargo.toml in /, it should
   work the same.

3. Record the SHA1 of HEAD in vendor/, to further reduce guesswork as to
   which baseline was in play when the vendor directory was last
   modified.

4. Do a dry-run update first, to expose what *could* be updated.

+++ maint/vendor_update.sh
@@ -1,12 +1,19 @@
+if [[ -z "$1" ]]; then
+  echo "pass a package to update dummy"
+  exit 1
+fi
 if [[ -e vendor/Cargo.lock ]]; then
-  rm -v vendor/Cargo.lock || exit 1
+  mv -v vendor/Cargo.lock ./Cargo.lock || exit 1
 fi
 if [[ -e vendor/config ]]; then
   rm -v vendor/config || exit 1
 fi
-cargo update || exit 1
+cargo update -v --dry-run || exit 1
+cargo update -v -p "$1" || exit 1
 cargo vendor --versioned-dirs || exit 1
 cp -v Cargo.lock vendor/ || exit 1
+cp -v Cargo.toml vendor/Cargo.toml.orig || exit  1
+git rev-parse HEAD > vendor/master-commit || exit 1
 cat <<"EOF" > vendor/config
 [source.crates-io]
 replace-with = "vendored-sources"
